Extract functions to utils.py
